Upgraded ID3Bot.py card choosing
