+++ Uno/DecisionTrees/C4_5Tree.py
@@ -35,6 +35,7 @@ class C4_5Tree:
         self.node_value = node_value
         self.split_attribute = split_attribute
         self.column_names = column_names
+        self.prediction_data = None
 
     def get_labels(self):
         return self.labels_encodes
@@ -194,6 +195,16 @@ class C4_5Tree:
 
         return output_series
 
+    def assign_prediction_data(self):
+        self.prediction_data = pd.Series(data=self.Y_data[self.remaining_data_indices]).replace(
+            self.labels_encodes["card_played"].keys(), self.labels_encodes["card_played"].values()).value_counts(sort=True)
+
+
+        if self.is_leaf:
+            return
+
+        for child in self.children:
+            child.assign_prediction_data()
 
     def sort_children(self):
         self.children.sort(key=lambda x: x.node_value, reverse=False)
@@ -231,6 +242,18 @@ class C4_5Tree:
             spaces += "|     "
         print(f"{spaces}{self.node_depth}. {self.node_value}: {list(self.labels_encodes.keys())[self.split_attribute]} -> len: {self.remaining_data_indices.shape[0]}")
 
+    def print_tree(self):
+        str = ""
+        for i in range(0, self.node_depth):
+            str += f"\t"
+
+        if self.parent == None:
+            print(f"{str}{self.node_depth}. {self.node_value}")
+        else:
+            print(f"{str}{self.node_depth}. p_s={self.parent.split_attribute} -> n_v={self.node_value}")
+
+        for child in self.children:
+            child.print_tree()
 
 def load_tree(filename):
     abs_path = os.path.abspath(filename)
@@ -240,29 +263,27 @@ def load_tree(filename):
 
 
 def main():
-    df = pd.read_csv("Uno/games_data/MergedCSV/20240728_2356_uno_game_693MB_testing.csv")
+    df = pd.read_csv("Uno/games_data/MergedCSV/20241226_1734_uno_game.csv")
     df = prepare_data_for_learning(df)
-    df = df.loc[df["card_played"] != "Draw"]
-    df = df.loc[df["card_played"] != "All Colors"]
-    df = df.loc[df["card_played"] != "+4 Colors"]
-    new_df = pd.DataFrame(columns=df.columns)
-    for card in df["card_played"].value_counts().index:
-        new_df = pd.concat([new_df, df[df["card_played"] == card].iloc[:9780]], ignore_index=True)
-    print(new_df["card_played"].value_counts())
-    print(f"Data loaded...: \n {df["card_played"].value_counts()}")
-    df, label_encoders = encode_data(new_df)
+    print(df.head())
+    # new_df = pd.DataFrame(columns=df.columns)
+    # print(new_df["card_played"].value_counts())
+    # print(f"Data loaded...: \n {df["card_played"].value_counts()}")
+    df, label_encoders = encode_data(df)
     X_data = df.iloc[:, :-1].to_numpy()
     Y_data = df.iloc[:, -1].to_numpy()
     start = time.time()
     tree = C4_5Tree(X_data=X_data, Y_data=Y_data, remaining_data_indices=df.index.to_numpy(),
                     labels_encodes=label_encoders, column_names=df.columns.to_list()[:-1])
-    max_depth = 25
-    min_values_in_leaf = 1000
-    min_gain_ratio = 0.4
-    tree.build_tree(max_depth=25, min_values_per_leaf=1000, min_gain_ratio=0.04)
+    max_depth = 100
+    min_values_in_leaf = 200
+    min_gain_ratio = 0.03
+    tree.build_tree(max_depth=max_depth, min_values_per_leaf=min_values_in_leaf, min_gain_ratio=min_gain_ratio)
     print(f"Tree successfully built in {time.time() - start} seconds.")
     tree.sort_children()
-    tree.save_tree(f"C4_5Tree_tree_d{max_depth}_mvl{min_values_in_leaf}_gr{min_gain_ratio}.pkl")
+    tree.print_tree()
+    tree.assign_prediction_data()
+    tree.save_tree(f"3GB_Dataset_C4_5Tree_tree_d{max_depth}_mvl{min_values_in_leaf}_gr{min_gain_ratio}.pkl")
 
 if __name__ == "__main__":
     main()
